pyproject.toml
[tool.poetry]
name = "repository-pattern"
version = "0.1.0"
description = ""
authors = ["victor-pi-almeida <victorres.almeida@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
SQLAlchemy = "^2.0.31"
alembic = "^1.13.2"
pydantic = "^2.8.2"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
pytest-mock = "^3.14.0"
psycopg2 = "^2.9.9"
pytz = "^2024.1"
python-dateutil = "^2.9.0.post0"
pylint-flask = "^0.6"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.3.2"
flake8 = "^7.1.0"
pylint = "^3.2.6"
pylint-flask = "^0.6"
tox = "^4.16.0"
coverage = "^7.6.0"
pytest-cov = "^5.0.0"
pytest-flask = "^1.3.0"
flask-testing = "^0.8.1"


[tool.flake8]
max-line-length = 80
select = "F,E,W,B,B901,B902,B903"
exclude = [
    ".git",
    "packages",
]
ignore = [
    "E722",
    "B001",
    "W503",
    "E203",
    "E402"
]

[tool.pylint]
max-line-length = 80
init-hook = "import sys; sys.path.append('.')"


[tool.pylint.'FORMAT']
max-line-length=120

[tool.pylint.messages_control]
disable = "w0012, R0903, W0107, C0116, C0415, W0621, W0622, W0212"

[tool.black]
line-length = 120

[tool.pytest]
paths = ["tests/"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
